refactor: reorganize authentication classes, move UserAuthenticationProvider and UserAuthenticationEntryPoint to security package
