+++ backend/src/main/java/ch/sectioninformatique/template/security/JwtAuthFilter.java
@@ -11,8 +11,6 @@ import org.springframework.web.filter.OncePerRequestFilter;
 import org.springframework.stereotype.Component;
 import org.springframework.lang.NonNull;
 
-import ch.sectioninformatique.template.user.UserAuthenticationProvider;
-
 import java.io.IOException;
 
 /**
