+++ backend/src/test/java/ch/sectioninformatique/template/user/UserAuthenticationProviderTest.java
@@ -9,6 +9,8 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 
+import ch.sectioninformatique.template.security.UserAuthenticationProvider;
+
 import java.lang.reflect.Method;
 import java.util.Arrays;
 import java.util.List;
